% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineTable.R
\name{baselineTable}
\alias{baselineTable}
\title{A Baseline Table/ Descriptive Statistics function}
\usage{
baselineTable(
  data,
  vars,
  labels = NULL,
  grouping.var,
  round_dec = 2,
  print.vars = FALSE
)
}
\arguments{
\item{data}{Data needs to be entered that includes relevant variables for the baseline table}

\item{vars}{Vector of variables to be included as rows for baseline Table}

\item{labels}{Vector of labels to be set in rows for baseline Table. Needs to be the same length as vars. Defaults to vars values.}

\item{grouping.var}{Optional variable name if you want to include extra rows (and statistical tests) for baseline variables split by groups.}

\item{round_dec}{Number of decimal spaced to-be-included in baseline Table. Default is 2}

\item{print.vars}{Should variable names be included in the output (in addition to labels)? Default is FALSE.}
}
\description{
This function creates a baseline table including descriptive statistics and automates work for a "Table 1" in scientific publications.
}
\examples{
baselineTable()
}
\keyword{Baseline}
\keyword{Table}
