% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertORSMD.R
\name{convertORSMD}
\alias{convertORSMD}
\title{Conversion between Odds Ratios (OR) and Standardised Mean Differences (SMD)}
\usage{
convertORSMD(input, initialMetric = "OR")
}
\arguments{
\item{input}{Input needs to be a value/vector of ORs or SMDs. Type of metric is defined in the initialMetric argument.}

\item{initialMetric}{Initial metric provided to input argument needs to be defined as "OR" or "SMD". Default is "OR".}
}
\description{
This function converts ORs to SMDs or vice versa using the formula SMD=log(OR)/1.81, which was proposed by Chinn et al. (2000).
}
\examples{
convertORSMD(input = 1.2, initialMetric = "OR")
convertORSMD(input = -0.2, initialMetric = "SMD")
}
\references{
Chinn S. A simple method for converting an odds ratio to effect size for use in meta-analysis. Stat Med 2000; 19: 3127–3131.
}
\author{
Nils Kappelmann
}
\keyword{Effect}
\keyword{OR}
\keyword{SMD,}
\keyword{conversion;}
\keyword{size}
