% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createOutputTable.R
\name{createOutputTable}
\alias{createOutputTable}
\title{A helper function for baselineTable to create the general output table}
\usage{
createOutputTable(data, vars, labels, var.details, round_dec)
}
\arguments{
\item{var.details}{This additional argument specifies nature of variables (i.e., continuous or categorical). For detailed documentation, see the original code from baselineTable.}

\item{...}{It takes arguments from baselineTable}
}
\description{
This function is called within baselineTable and creates additional rows for each level specified in grouping.var. It also uses standard statistical tests to compare descriptive statistics between groups.
}
\examples{
createOutputTable()
}
\keyword{Baseline}
\keyword{Table}
