% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportPubHelpercsv.R
\name{exportPubHelpercsv}
\alias{exportPubHelpercsv}
\title{exportPubHelpercsv}
\usage{
exportPubHelpercsv(object = NULL, file = NULL, sep = ",")
}
\arguments{
\item{object}{Output data.frame for export as .csv file.}

\item{file}{Filename (and location).}

\item{sep}{Separator for csv file. Defaults is ",".}
}
\description{
This function is a short pre-specified version of write.table with options indicated for table inclusion in scientific publications.
}
\examples{
data(airquality)
model = lm(Temp ~ Wind, data = airquality)
outputTable = formatGLMTable(model = model)
exportPubHelpercsv(outputTable, file = "PubHelperTest.csv")
}
\author{
Nils Kappelmann
}
\keyword{csv;}
\keyword{export}
