% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGLMTable.R
\name{formatGLMTable}
\alias{formatGLMTable}
\title{Formatting GLM results for table export}
\usage{
formatGLMTable(model = NULL, round_dec = 2, lm.ci = FALSE, ...)
}
\arguments{
\item{model}{Data needs to be entered that includes relevant variables for the baseline table}

\item{round_dec}{Number of decimal spaces to-be-included in baseline Table. Default is 2}

\item{lm.ci}{Indicate if 95\% confidence interval from LM should be included in output. Default is FALSE.}

\item{...}{Other arguments from getGLMTable}
}
\description{
This function creates a formatted table including output from common GLM models.
}
\examples{
data(airquality)
model = lm(Temp ~ Wind, data = airquality)
formatGLMTable(model = model)
}
\author{
Nils Kappelmann
}
\keyword{GLM;}
\keyword{table}
