% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGLMTables.R
\name{mapGLMTables}
\alias{mapGLMTables}
\title{Loop over getGLMTable}
\usage{
mapGLMTables(
  data = NULL,
  y = NULL,
  x = NULL,
  z = NULL,
  model.type = "lm",
  simplify = TRUE
)
}
\arguments{
\item{data}{Data.frame including model variables}

\item{y}{Vector of outcome variables}

\item{x}{Vector of predictor variables}

\item{z}{Vector of covariates if covariates are included}

\item{model.type}{Specify which statistical model to run. Options are "lm" for linear regression, "glm" for logistic regression, and "polr" for ordinal logistic regression}

\item{simplify}{Should full lm models and GLMTables be excluded from output? Default is TRUE.}
}
\description{
This function is a wrapper that loops over getGLMTable or formatGLMTable to create a single output data frame
}
\examples{
data(airquality)
mapGLMTables(data = airquality, y = "Ozone", x = c("Solar.R", "Wind"), z = "Temp")
}
\author{
Nils Kappelmann
}
\keyword{GLM;}
\keyword{table}
