% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanpooled.R
\name{meanpooled}
\alias{meanpooled}
\title{Calculation of Pooled Mean}
\usage{
meanpooled(n1 = NA, n2 = NA, mean1 = NA, mean2 = NA)
}
\arguments{
\item{n1}{Sample size of sample/group 1.}

\item{n2}{Sample size of sample/group 2.}

\item{mean1}{Mean of sample/group 1.}

\item{mean2}{Mean of sample/group 2.}
}
\description{
This function computes the pooled mean from two samples/groups.
}
\examples{
meanpooled(n1 = 30, n2 = 20, mean1 = 15, mean2 = 20)
}
\author{
Nils Kappelmann
}
\keyword{mean}
