% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SDfromCI.R
\name{SDfromCI}
\alias{SDfromCI}
\title{Inference of Standard Deviation (SD) from Confidence Interval (CI)}
\usage{
SDfromCI(n = NA, ci.lb = NA, ci.ub = NA)
}
\arguments{
\item{n}{Sample size.}

\item{ci.lb}{Lower bound of 95\% CI.}

\item{ci.ub}{Upper bound of 95\% CI.}
}
\description{
This function infers the SD from a given 95\% CI.
}
\examples{
SDfromCI(n = 30, ci.lb = -0.1, ci.ub = 0.2)
}
\author{
Nils Kappelmann
}
\keyword{confidence}
\keyword{deviation;}
\keyword{interval}
\keyword{standard}
