% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sdpooled.R
\name{sdpooled}
\alias{sdpooled}
\title{Calculation of Pooled Standard Deviation (SD)}
\usage{
sdpooled(n1 = NA, n2 = NA, sd1 = NA, sd2 = NA)
}
\arguments{
\item{n1}{Sample size of sample/group 1.}

\item{n2}{Sample size of sample/group 2.}

\item{sd1}{SD of sample/group 1.}

\item{sd2}{SD of sample/group 2.}
}
\description{
This function computes the pooled SD from two samples/groups.
}
\examples{
sdpooled(n1 = 30, n2 = 20, sd1 = 15, sd2 = 20)
}
\author{
Nils Kappelmann
}
\keyword{deviation}
\keyword{standard}
