% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addGroupInfoToTable.R
\name{addGroupInfoToTable}
\alias{addGroupInfoToTable}
\title{A helper function for baselineTable to specify additional rows split by Group}
\usage{
addGroupInfoToTable(data, output, vars, grouping.var)
}
\arguments{
\item{...}{It takes arguments from baselineTable}
}
\description{
This function is called within baselineTable and creates additional rows for each level specified in grouping.var. It also uses standard statistical tests to compare descriptive statistics between groups.
}
\examples{
addGroupInfoToTable()
}
\keyword{Baseline}
\keyword{Table}
