% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/baselineTable.R
\name{baselineTable}
\alias{baselineTable}
\title{A Baseline Table/ Descriptive Statistics function}
\usage{
baselineTable(
  data,
  vars,
  labels = NULL,
  grouping.var = NULL,
  round_dec = 2,
  placeholder = "   ",
  welch.test = FALSE,
  na.statistics = "all",
  print.vars = FALSE
)
}
\arguments{
\item{data}{Data needs to be entered that includes relevant variables for the baseline table}

\item{vars}{Vector of variables to be included as rows for baseline Table}

\item{labels}{Vector of labels to be set in rows for baseline Table. Needs to be the same length as vars. Defaults to vars values.}

\item{grouping.var}{Optional variable name if you want to include extra rows (and statistical tests) for baseline variables split by groups.}

\item{round_dec}{Number of decimal spaces to-be-included in baseline Table. Default is 2}

\item{placeholder}{The placeholder can be specified with indents descriptive statistics labels (e.g., Mean (SD)) by a given string. The default is "   " (i.e., three spaces).}

\item{welch.test}{If grouping.var is specified and includes two groups only, The Welch test will be used to test for significant differences between groups on continuous variables. Default is FALSE.}

\item{na.statistics}{Should Missing value estimates be included in the output? Default is "all"; i.e., to provide all missingness statistics. Other options include "if.applicable" and "none". "if.applicable" includes missingness statistics for those variables, where NAs are present.}

\item{print.vars}{Should variable names be included in the output (in addition to labels)? Default is FALSE.}
}
\description{
This function creates a baseline table including descriptive statistics and automates work for a "Table 1" in scientific publications.
}
\examples{
data("airquality")
airquality$Period = ifelse(airquality$Month \%in\% 5:7, "Early Summer", "Late Summer")

baselineTable(data = airquality,
vars = c("Ozone", "Solar.R",
"Wind", "Temp", "Period"))
baselineTable(data = airquality,
vars = c("Ozone", "Solar.R",
"Wind", "Temp", "Period"),
labels = c("Ozone (ppb)", "Solar R (lang)",
"Wind (mph)", "Temperature (degrees F)", "Period"))
baselineTable(data = airquality,
vars = c("Ozone", "Solar.R", "Wind", "Temp"),
labels = c("Ozone (ppb)", "Solar R (lang)",
"Wind (mph)", "Temperature (degrees F)"),
grouping.var = "Period", round_dec = 1)
}
\author{
Nils Kappelmann
}
\keyword{Baseline}
\keyword{Table}
