% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formatGLMTable.R, R/getGLMTable.R
\name{formatGLMTable}
\alias{formatGLMTable}
\title{Formatting GLM results for table export}
\usage{
formatGLMTable(model = NULL)

formatGLMTable(model = NULL)
}
\arguments{
\item{model}{Data needs to be entered that includes relevant variables for the baseline table}

\item{intercept}{Should intercepts be included in the output? Default is TRUE}

\item{exclude.covariates}{Specify covariates that should be excluded from the output.}

\item{round_dec}{Number of decimal spaces to-be-included in baseline Table. Default is 2}

\item{lm.ci}{Indicate if 95\% confidence interval from LM should be included in output. Default is FALSE.}
}
\description{
This function creates a formatted table including output from common GLM models.

This function creates a table including output from common GLM models. formatGLMtable offers further formatting for direct inclusion in scientific publications
}
\examples{
data(airquality)
model = lm(Temp ~ Wind, data = airquality)
formatGLMTable(model = model)
data(airquality)
model = lm(Temp ~ Wind, data = airquality)
getGLMTable(model = model)
}
\author{
Nils Kappelmann

Nils Kappelmann
}
\keyword{GLM;}
\keyword{table}
