% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGLMTable.R
\name{getGLMTable}
\alias{getGLMTable}
\title{Get a Table of GLM results}
\usage{
getGLMTable(model = NULL)
}
\arguments{
\item{model}{Data needs to be entered that includes relevant variables for the baseline table}
}
\description{
This function creates a table including output from common GLM models. formatGLMtable offers further formatting for direct inclusion in scientific publications
}
\examples{
data(airquality)
model = lm(Temp ~ Wind, data = airquality)
getGLMTable(model = model)
}
\author{
Nils Kappelmann
}
\keyword{GLM;}
\keyword{table}
