% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapGLMTables.R
\name{mapGLMTables}
\alias{mapGLMTables}
\title{mapGLMTables: Loop over regression models and extract results.}
\usage{
mapGLMTables(
  data = NULL,
  y = NULL,
  x = NULL,
  z = NULL,
  model.type = "lm",
  format.output = FALSE,
  simplify = TRUE
)
}
\arguments{
\item{data}{Data.frame including model variables}

\item{y}{Vector of outcome variables}

\item{x}{Vector of predictor variables}

\item{z}{Vector of covariates if covariates are included}

\item{model.type}{Specify which statistical model to run. Options are "lm" for linear regression, "glm" for logistic regression, and "polr" for ordinal logistic regression}

\item{format.output}{Should output be formatted for scientific publications? Default is FALSE.}

\item{simplify}{Should results be simplified to relevant regression parameters of the predictor? If not, full lm models, GLMTables, and plots to check model assumptions (using the performance package) are included in the output. Default is simplify=TRUE.}
}
\description{
mapGLMTables loops over regression models (linear, logistic, & ordinal logistic) and returns summarised output from getGLMTable. It is possible to include covariates, which are automatically excluded from the output. However, the full output can be obtained as a more complex output by setting simplify=FALSE. Note that a complex model output is returned if predictor variables include factors with >2 levels.
}
\examples{
data(airquality)
mapGLMTables(data = airquality, y = "Ozone", x = c("Solar.R", "Wind"), z = "Temp")
}
\author{
Nils Kappelmann
}
\keyword{GLM;}
\keyword{table}
